Refactor configuration interfaces: move v10 config interfaces to a new file and update imports
