+++ src/app/core/interfaces/v10-config-interface.ts
