+++ src/app/widgets/widget-label/widget-label.component.scss
@@ -5,12 +5,6 @@
   overflow: hidden;
 }
 
-.label-wrapper {
-  width: 100%;
-  height: 100%;
-  min-height: 1px;
-}
-
 .label-canvas {
   width: 100% !important;
   height: 100% !important;
